Reduce reliance on RenderingDevice::GetCamera() since it can switch projection when the render target changes. This causes issues if any code in ProcessMessage/AdvanceTime makes use of the camera to determine visibility of objects (i.e. particle systems). Instead, use GameView (global 'gameView') to keep a camera that always represents the view/projection of the actual game view, which is never touched.

+++ Infrastructure/src/graphics/camera.cpp
@@ -191,8 +191,8 @@ namespace gfx {
 	XMFLOAT3 WorldCamera::TileToWorld(locXY tilePos)
 	{
 		auto result = XMFLOAT3();
-		result.x = (tilePos.locy - tilePos.locx - 1) * 20;
-		result.y = (tilePos.locy + tilePos.locx ) * 14;
+		result.x = (float)((tilePos.locy - tilePos.locx - 1) * 20);
+		result.y = (float)((tilePos.locy + tilePos.locx) * 14);
 		result.z = 0;
 
 		return result;
